Added quest functionality
GetOrderSentences - Display all sentences of the given text in ascending order of the number of words in each of them.
GetWordsFromQuestions - Find and print words of a given length in all interrogative sentences of the text.
DeleteWords - Remove from the text all words of a given length beginning with a consonant letter.